+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [2.3.0] 2025-11-04 More XWN support
+
+- Custom currency system configuration supported. Base currency and up to 5 custom currencies.
 - Added readied and stowed modifiers to character sheet (under features/tweaks)
 - Fixed capt support department bug